+++ examples/apps/branching/src/groceryList/index.ts
@@ -1,0 +1,10 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export {
+	applyDiffToGroceryList,
+	diffGroceryListJSON,
+	GroceryListFactory,
+} from "./groceryList.js";
