--- examples/apps/branching/src/model/index.ts
@@ -1,7 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export { GroceryListContainerRuntimeFactory } from "./containerCode.js";
-export { applyDiffToGroceryList, diffGroceryListJSON } from "./groceryList.js";
